refactor: consolidate HubTitle tests into Canvas stories following new guidelines
- Remove HubTitle.stories.tsx following new development guideline:
  "基本的にメインコンポーネントのみstoryファイルを作り、サブコンポーネントは作成しない"
- Integrate HubTitle tests into Canvas.stories.tsx as sub-component tests:
  - Add HubTitleAccessibilityTest: accessibility requirements testing
  - Add HubTitleCustomStyleTest: custom style application verification
  - Add HubTitleStateChangeTest: Hub state change and dynamic title updates with useHubStore
- Update Canvas component description to note HubTitle tests inclusion
- All 30 Storybook tests continue to pass after consolidation
- Reduce test count from 33 to 30 while maintaining comprehensive coverage
