+++ docs/03.developmentAgreement.md
@@ -279,6 +279,7 @@ src/
 - **ストーリーは「生きた仕様書」**: 各コンポーネントの「状態」と「振る舞い」を網羅するカタログとして機能させる。
 - **効率重視**: 個人開発の速度を落とさないよう、規約はシンプルに保ち、自動化できる部分は積極的に活用する。
 - **関心の分離**: ストーリーファイル（`.stories.tsx`）は、あくまでコンポーネントの **「見せ方」と「使い方」** の定義に集中する。複雑なロジックはコンポーネントやカスタムフック内に閉じ込める。
+- 基本的にメインコンポーネントのみstoryファイルを作り、サブコンポーネントは作成しない
 
 #### 5.3.2. CSF (Component Story Format) 3.0 規約
