+++ src/components/layouts/Canvas/__stories__/Canvas.stories.tsx
@@ -1,6 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { within, expect } from '@storybook/test';
+import { within, expect, userEvent } from '@storybook/test';
 import { Canvas } from '../components/Canvas';
+import { useHubStore } from '@/stores/hubStore';
 
 const meta: Meta<typeof Canvas> = {
   title: 'Components/Layouts/Canvas',
@@ -13,7 +14,7 @@ const meta: Meta<typeof Canvas> = {
     docs: {
       description: {
         component:
-          'インタラクティブ・ハブ画面のメイン表示エリア（Canvas）コンポーネント。HubTitleとCircleOfFifthsを統合し、Hub状態管理（useHubStore）と連携して動的なタイトル表示を提供します。現在は五度圏を表示し、将来的にはクロマチックサークルとの切り替えにも対応します。',
+          'インタラクティブ・ハブ画面のメイン表示エリア（Canvas）コンポーネント。HubTitleとCircleOfFifthsを統合し、Hub状態管理（useHubStore）と連携して動的なタイトル表示を提供します。現在は五度圏を表示し、将来的にはクロマチックサークルとの切り替えにも対応します。注意：HubTitle（サブコンポーネント）のテストもこのストーリーに含まれています。',
       },
     },
     a11y: {
@@ -184,5 +185,140 @@ export const CustomStyle: Story = {
         story: 'カスタムスタイルを適用したCanvasです。背景色やボーダーをカスタマイズできます。',
       },
     },
+  },
+};
+
+/**
+ * HubTitleのアクセシビリティテスト
+ * HubTitle（サブコンポーネント）のアクセシビリティ要件をCanvasから確認
+ */
+export const HubTitleAccessibilityTest: Story = {
+  args: {},
+  parameters: {
+    docs: {
+      description: {
+        story:
+          'Canvas内のHubTitleコンポーネントのアクセシビリティ要件をテストします。適切な見出しレベル、セマンティクスを確認します。',
+      },
+    },
+  },
+  play: async ({ canvasElement }) => {
+    const canvas = within(canvasElement);
+
+    // 見出しの階層構造確認
+    const heading = canvas.getByRole('heading', { level: 1 });
+    expect(heading).toBeInTheDocument();
+
+    // テキストコンテンツの確認
+    expect(heading).toHaveTextContent(/^(Circle of Fifths|Chromatic Circle)$/);
+
+    // 基本的なCSSクラスの確認
+    expect(heading).toHaveClass('text-title');
+  },
+};
+
+/**
+ * HubTitleのカスタムスタイルテスト
+ * HubTitle（サブコンポーネント）のスタイルカスタマイズをCanvasから確認
+ */
+export const HubTitleCustomStyleTest: Story = {
+  args: {},
+  parameters: {
+    docs: {
+      description: {
+        story:
+          'Canvas内のHubTitleコンポーネントにカスタムスタイルが適用されていることを確認します。',
+      },
+    },
+  },
+  play: async ({ canvasElement }) => {
+    const canvas = within(canvasElement);
+
+    // HubTitleの表示確認
+    const hubTitle = canvas.getByRole('heading', { level: 1 });
+    expect(hubTitle).toBeInTheDocument();
+
+    // Canvas実装で適用されているカスタムクラスの確認
+    expect(hubTitle).toHaveClass('text-center', 'text-2xl', 'lg:text-4xl');
+
+    // セマンティクスの確認
+    expect(hubTitle).toHaveClass('text-title');
+  },
+};
+
+/**
+ * Hub状態変更とHubTitleの連携テスト
+ * useHubStoreとの連携によるHubTitleの動的変更をCanvasから確認
+ */
+export const HubTitleStateChangeTest: Story = {
+  args: {},
+  parameters: {
+    docs: {
+      description: {
+        story:
+          'Hub種類切り替え機能とCanvas内HubTitleの連携テストです。状態管理でタイトルが正しく変更されることを確認します。',
+      },
+    },
+  },
+  decorators: [
+    Story => {
+      const { hubType, setHubType } = useHubStore();
+
+      return (
+        <div className="min-h-screen w-full bg-gradient-to-b from-gray-900 to-black text-white">
+          <Story />
+          {/* テスト用コントロールボタン */}
+          <div className="absolute bottom-8 left-1/2 -translate-x-1/2 transform">
+            <div className="flex space-x-4">
+              <button
+                data-testid="circle-of-fifths-button"
+                onClick={() => setHubType('circle-of-fifths')}
+                className={`rounded-md px-4 py-2 font-medium transition-colors ${
+                  hubType === 'circle-of-fifths'
+                    ? 'bg-blue-600 text-white'
+                    : 'bg-gray-200 text-gray-700 hover:bg-gray-300'
+                }`}
+              >
+                五度圏
+              </button>
+              <button
+                data-testid="chromatic-circle-button"
+                onClick={() => setHubType('chromatic-circle')}
+                className={`rounded-md px-4 py-2 font-medium transition-colors ${
+                  hubType === 'chromatic-circle'
+                    ? 'bg-blue-600 text-white'
+                    : 'bg-gray-200 text-gray-700 hover:bg-gray-300'
+                }`}
+              >
+                クロマチックサークル
+              </button>
+            </div>
+          </div>
+        </div>
+      );
+    },
+  ],
+  play: async ({ canvasElement }) => {
+    const canvas = within(canvasElement);
+
+    // テスト開始前にストアを初期状態に確実にセット
+    useHubStore.setState({ hubType: 'circle-of-fifths' });
+
+    // 初期状態でh1タグに「Circle of Fifths」が表示されていることを確認
+    expect(canvas.getByRole('heading', { name: 'Circle of Fifths' })).toBeInTheDocument();
+
+    // クロマチックサークルボタンをクリック
+    const chromaticButton = canvas.getByTestId('chromatic-circle-button');
+    await userEvent.click(chromaticButton);
+
+    // タイトルが「Chromatic Circle」に変更されることを確認
+    expect(canvas.getByRole('heading', { name: 'Chromatic Circle' })).toBeInTheDocument();
+
+    // 五度圏ボタンをクリック
+    const circleButton = canvas.getByTestId('circle-of-fifths-button');
+    await userEvent.click(circleButton);
+
+    // タイトルが「Circle of Fifths」に戻ることを確認
+    expect(canvas.getByRole('heading', { name: 'Circle of Fifths' })).toBeInTheDocument();
   },
 };
